+++ tests/Bot/ProviderTest.php
@@ -6,6 +6,7 @@ use Mockery;
 use PHPUnit_Framework_TestCase;
 use seregazhuk\PinterestBot\Api\Request;
 use seregazhuk\PinterestBot\Api\Response;
+use seregazhuk\tests\Helpers\ResponseHelper;
 use seregazhuk\PinterestBot\Api\Providers\Core\Provider;
 
 /**
@@ -13,6 +14,7 @@ use seregazhuk\PinterestBot\Api\Providers\Core\Provider;
  */
 class ProviderTest extends PHPUnit_Framework_TestCase
 {
+    use ResponseHelper;
 
     protected function tearDown()
     {
@@ -20,7 +22,7 @@ class ProviderTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function it_returns_data_for_response_without_pagination()
+    public function it_returns_data_for_response()
     {
         $response = ['resource_response' => ['data' => 'value']];
 
@@ -32,27 +34,103 @@ class ProviderTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function it_returns_response_object_if_it_has_pagination()
+    public function it_should_return_response_object_for_pagination()
     {
-        $response = ['resource' => ['options' => ['bookmarks' => 'bookmarks_string']]];
+        $paginatedResponse = $this->createPaginatedResponse($this->paginatedResponse);
+
+        $request = $this->makeRequest($paginatedResponse, 2);
+        $request
+            ->shouldReceive('exec')
+            ->once()
+            ->andReturn(json_encode([]));
+
+        /** @var DummyProvider $provider */
+        $provider = Mockery::mock(DummyProvider::class, [$request, new Response()])
+            ->makePartial();
+
+        $provider->dummyPaginate(['test' => 'test'], 'http://example.com')->toArray();
+    }
+
+    /** @test */
+    public function it_should_clear_response_before_pagination()
+    {
+        $response = Mockery::mock(Response::class)
+            ->shouldReceive('clear')
+            ->once()
+            ->getMock()
+            ->makePartial();
+
+        /** @var DummyProvider $provider */
+        $provider = Mockery::mock(DummyProvider::class, [$this->makeRequest([]), $response])
+            ->makePartial();
+
+        $provider->dummyPaginate(['test' => 'test'], 'http://example.com')->toArray();
+    }
+
+
+    /** @test */
+    public function it_should_return_bool_if_required_for_post_request()
+    {
+        $response = ['resource_response' => ['data' => 'value']];
 
         $provider = $this->makeProvider($response);
 
-        $this->assertInstanceOf(Response::class, $provider->visitPage());
+        $this->assertTrue($provider->dummyPost());
     }
 
     /**
-     * @param $response
-     * @return Mockery\Mock|Provider
+     * @param mixed $response
+     * @param int $times
+     * @return Mockery\Mock|Provider|DummyProvider
      */
-    protected function makeProvider($response)
+    protected function makeProvider($response, $times = 1)
     {
-        $request = Mockery::mock(Request::class)
+        $request = $this->makeRequest($response, $times);
+
+        return Mockery::mock(DummyProvider::class, [$request, new Response()])
+            ->makePartial();
+    }
+
+    /**
+     * @param mixed $response
+     * @param int $times
+     * @return Mockery\MockInterface
+     */
+    protected function makeRequest($response, $times = 1)
+    {
+        return Mockery::mock(Request::class)
             ->shouldReceive('exec')
+            ->times($times)
             ->andReturn(json_encode($response))
             ->getMock();
+    }
+}
 
-        return Mockery::mock(Provider::class, [$request, new Response()])
-            ->makePartial();
+class DummyProvider extends Provider {
+
+    /**
+     * @param mixed $data
+     * @param string $resourceUrl
+     * @return \seregazhuk\PinterestBot\Helpers\Pagination
+     */
+    public function dummyPaginate($data, $resourceUrl)
+    {
+        return $this->paginate($data, $resourceUrl);
+    }
+
+    /**
+     * @return array|bool|Response
+     */
+    public function dummyGet()
+    {
+        return $this->get([], '');
+    }
+
+    /**
+     * @return array|bool|Response
+     */
+    public function dummyPost()
+    {
+        return $this->post([], '');
     }
 }